refactor: remove header
